requirements.txt
fastapi
uvicorn
pydantic
supabase
python-dotenv
streamlit
pypdf
requests
httpx
openai
